[roottest] remove unused code for io/perf, add few more tests
SergeiShortTest and SergeiHardTest not included - while consume a lot of time
