--- roottest/root/io/perf/.rootrc
@@ -1,1 +1,0 @@
-Rint.History:            .root_hist
