--- roottest/root/io/perf/Makefile
@@ -1,34 +1,0 @@
-# This is a template for all makefile.
-
-#Set the list of files to be delete by clean:
-CLEAN_TARGETS += $(ALL_LIBRARIES)
-
-#Set the list of target to make while testing
-TEST_TARGETS += mytest
-
-# adjust the location of Rules.mk as needed.
-ifeq ($(strip $(ROOTTEST_HOME)),)
-   export ROOTTEST_HOME := $(shell git rev-parse --show-toplevel)/roottest/
-   ifeq ($(strip $(ROOTTEST_HOME)),)
-      export ROOTTEST_HOME := $(shell expr $(CURDIR) : '\(.*/roottest/\)')
-   endif
-   ifeq ($(strip $(ROOTTEST_HOME)),)
-      $(error The head of roottest was not found.  Set ROOTTEST_HOME)
-   endif
-endif
-
-include $(ROOTTEST_HOME)/scripts/Rules.mk
-
-
-mytest:
-	$(CMDECHO) echo mytest > /dev/null
-
-testWithFailure:
-ifeq ($(FAIL),)
-	$(WarnFailTest)
-endif
-
-perftest:
-	$(CMDECHO) $(CALLROOTEXE) -q julius.C+
-
-
