--- roottest/root/io/perf/orrun.C
@@ -1,4 +1,0 @@
-{
-gROOT->ProcessLine(".L OldReadBuffer_C.so");
-gROOT->Time();
-}
