--- roottest/root/io/perf/rrun.C
@@ -1,4 +1,0 @@
-{
-gROOT->ProcessLine(".L ReadBuffer.C+");
-gROOT->Time();
-}
