--- roottest/root/io/perf/setup.C
@@ -1,29 +1,0 @@
-{
-gROOT->Time(1);
-gROOT->ProcessLine(".L WriteBuffer.C+");
-}
-
-/* notes:
-
-InfoWriteBuffer(10000000)
-4: 4.840           40.520->40.560  
-3: 4.510  37.240->37.330 
-2: 4.710       39.100->39.340        => 37.760->37.940
-1: 4.590  37.820->38.430
-0: 4.630  37.840
--1        37.380->37.480
-
-InfoWriteBufferMix(100000000)
-
-InfoWriteBufferMix(10000000)
--1:  8.100->8.210
- 0: 10.370->10.520
- 1:  8.980->9.050
-1b:  8.670->8.710
-1c:  8.460->8.520
-1d:  8.090->8.190 : -> no loop, no init, no indirection
-
-3a:  8.800->8.830 
-3b:  8.390->8.530 
-3c:  7.450->7.600
-*/
